requirements.txt
aiogram>=3.4.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
cachetools>=5.3.0